os/board/rtl8730e: Fix init_timer -Wint-conversion compile warning
component/wifi/common/rtw_timer.c: In function 'init_timer':
component/wifi/common/rtw_timer.c:107:10: warning: passing argument 3 of 'rtos_timer_create_static' makes integer from pointer without a cast [-Wint-conversion]
  107 |          NULL, // Uniq id used to identify which timer expire..
      |          ^~~~
      |          |
      |          void *
In file included from /root/tizenrt/os/board/rtl8730e/src/component/os/os_wrapper/include/os_wrapper.h:206,
                 from component/wifi/common/rtw_timer.c:19:
/root/tizenrt/os/board/rtl8730e/src/component/os/os_wrapper/include/os_wrapper_timer.h:41:90: note: expected 'uint32_t' {aka 'unsigned int'} but argument is of type 'void *'
   41 | int rtos_timer_create_static(rtos_timer_t *pp_handle, const char *p_timer_name, uint32_t timer_id,
      |                                                                                 ~~~~~~~~~^~~~~~~~

+++ os/board/rtl8730e/src/component/wifi/common/rtw_timer.c
@@ -104,7 +104,7 @@ void init_timer(struct timer_list *timer, const char *name)
 	if (timer->timer_hdl == NULL) {
 		rtos_timer_create_static(&timer->timer_hdl,
 								 (const char *)name,	// Just a text name, not used by the RTOS kernel.
-								 NULL,	// Uniq id used to identify which timer expire..
+								 (uint32_t)NULL,	// Uniq id used to identify which timer expire..
 								 RTOS_MAX_DELAY, // Timer Period, not 0
 								 FALSE, // Whether timer will auto-load themselves when expires
 								 timer_wrapper); // Timer callback
